Add suggested slSerialize_unpacker names
These are generally accurate but there is room for some misstakes on which exact types are being read. Also not completely confident on the namespace. The naming is not official and follows some patterns: ReadOptional_ have a behavior of reading 1 byte and something else if the byte was not 0. ReadArray_ use a common pattern with a do while loop reading 1 byte, that many of something, and then repeating if the original byte was 255 to allow for more entries. There are some funcs reading array without this behavior and those should not use the same prefix to avoid confusion.